+++ src/Objects/Scene/ZCR_Mesh.cpp
@@ -53,10 +53,10 @@ void ZCR_Mesh::SwitchToCoordSystem(CoordSystem _coordSystem)
 ZC_uptr<ZC_RendererSet> ZCR_Mesh::MakeRendererSet(float totalLength)
 {
     //  (* 2 -> two vertices on one line) (+ 2 -> last line, if totalLength 10 than fifth line(center line) gonna be colored and 5 grey lines from sides
-    ZC_DA<ZC_Vec3<float>> vertices((totalLength * 2 + 2) * 2);  //  last (* 2) -> pluss same count perpendicular lines 
+    ZC_DA<ZC_Vec3<float>> vertices((static_cast<ulong>(totalLength) * 2 + 2) * 2);  //  last (* 2) -> pluss same count perpendicular lines 
     auto linePosition = -(totalLength / 2.f);
     auto axisLength = totalLength / 2.f;
-    size_t i = 0;
+    ulong i = 0;
     for ( ; i < vertices.size / 2; )
     {
         vertices[i++] = ZC_Vec3<float>(linePosition, axisLength, 0);
